docs: add Astro and SvelteKit MailChannels examples
