+++ examples/frameworks/astro/package.json
@@ -3,10 +3,14 @@
   "version": "0.0.0",
   "private": true,
   "type": "module",
+  "engines": {
+    "node": ">=22.12.0"
+  },
   "scripts": {
     "dev": "astro dev",
     "build": "astro build",
-    "start": "astro preview"
+    "start": "astro preview",
+    "astro": "astro"
   },
   "dependencies": {
     "@astrojs/node": "^10.0.6",
