+++ examples/frameworks/astro/src/layouts/Layout.astro
@@ -1,0 +1,5 @@
+---
+---
+
+<a href="/">Home</a>
+<slot />
